fixes crawler bug

+++ src/main/java/utils/crawler/SEFileCrawler.java
@@ -206,7 +206,7 @@ public class SEFileCrawler {
 
 			final long[] timestamps = new long[pfnsToCrawl.size()];
 
-			int totalPFNCount = 0, inaccessiblePFNs = 0, corruptPFNs = 0, okPFNs = 0, unknownStatusPFNs = 0;
+			long totalPFNCount = 0, inaccessiblePFNs = 0, corruptPFNs = 0, okPFNs = 0, unknownStatusPFNs = 0;
 			long fileSizeBytes = 0, downloadTotalDurationMillis = 0, downloadedPFNsTotalCount = 0;
 			long xrdfsTotalDurationMillis = 0, xrdfsPFNsTotalCount = 0;
